#259 - compareArrays now finds and reports moved entries.
The returned "edit script" is backwards compatible because moved entries still include "deleted" and "added" entries. To indicate where they moved to/from, they will now also have a "moved" property.

Added and deleted entries will also have an "idx" property that's needed for the move detection code. We could remove the property afterwards, but that would hurt performance a bit (and add code). The only problem with the extra property is the need to modify the specs since they don't tolerate extra properties.

One of the existing specs already has an entry that triggers the "moved" detection, so I modified it to match the new behavior. Also I added a spec to demonstrate how compareArrays handles an array changed to completely new items.